% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outrunPalettes.R
\name{outrunPalette}
\alias{outrunPalette}
\title{An outrun themed palette generator}
\usage{
outrunPalette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{The name of a color palette. Options are \code{pal1},
\code{pal2}, \code{pal3}, \code{pal4}.}

\item{n}{Number of desired colors. All palettes are 5 colors.}

\item{type}{Either "continuous" or "discrete". Use "continuous" if you want
to automatically interpolate between colors.}
}
\value{
A vector of colors
}
\description{
A selection of 4 outrun themed palettes taken from this
}
\examples{
outrunPalette("pal1")
outrunPalette("pal12", n = 3)
}
