% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outrunPalettes.R
\docType{data}
\name{outrunPalettes}
\alias{outrunPalettes}
\title{Complete list of palettes}
\format{
An object of class \code{list} of length 4.
}
\usage{
outrunPalettes
}
\description{
Use \code{\link{outrunPalette}} to construct palettes of desired length.
}
\keyword{datasets}
